requirements.txt
beautifulsoup4==4.14.2
chromadb==0.5.3
fastapi==0.121.2
google-adk==1.2.1
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-cloud-aiplatform==1.127.0
google-cloud-appengine-logging==1.7.0
google-cloud-audit-log==0.4.0
google-cloud-bigquery==3.38.0
google-cloud-core==2.5.0
google-cloud-logging==3.12.1
google-cloud-resource-manager==1.15.0
google-cloud-secret-manager==2.25.0
google-cloud-speech==2.34.0
google-cloud-storage==2.19.0
google-cloud-trace==1.17.0
google-crc32c==1.7.1
google-genai==1.50.1
google-resumable-media==2.7.2
googleapis-common-protos==1.72.0
jinja2==3.1.0
requests==2.32.5
uvicorn==0.38.0
websockets==15.0.1
zipp==3.23.0
sqlalchemy==2.0.44
